requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.4
urllib3==1.26.14
